Avoid race condition: d/l soljs in parent process
Previously download of soljs executable was happening in parallel in
child processes, risking a possible race condition.  With this change,
the download happens in the single-threaded parent process instead.
